+++ Modules/Intent.Modules.AspNetCore.HealthChecks/Templates/HealthChecksConfiguration/HealthChecksConfigurationTemplatePartial.cs
@@ -41,15 +41,11 @@ namespace Intent.Modules.AspNetCore.HealthChecks.Templates.HealthChecksConfigura
 
             if (ExecutionContext.Settings.GetHealthChecks().HealthChecksUI())
             {
-                if (OutputTarget.GetProject().GetMaxNetAppVersion().Major >= 8)
-                {
-                    Logging.Log.Warning("""
-                                        Please be aware of the following issues when using Health Checks UI on .NET 8 or later.
-                                        - [Icons missing after upgrading to v8.0.0](https://github.com/Xabaril/AspNetCore.Diagnostics.HealthChecks/issues/2130).
-                                        - [[UI] Relative Address for HealthCheckEndpoint with Kestrel at http://0.0.0.0:0](https://github.com/Xabaril/AspNetCore.Diagnostics.HealthChecks/issues/410).
-                                        """);
-                }
-
+                Logging.Log.Warning("""
+                                    Please be aware of the following issues when using Health Checks UI with Docker.
+                                    - [[UI] Relative Address for HealthCheckEndpoint with Kestrel at http://0.0.0.0:0](https://github.com/Xabaril/AspNetCore.Diagnostics.HealthChecks/issues/410).
+                                    """);
+                
                 AddNugetDependency(NugetPackages.AspNetCoreHealthChecksUI(outputTarget));
                 AddNugetDependency(NugetPackages.AspNetCoreHealthChecksUIInMemoryStorage(outputTarget));
             }