+++ Tests/AdvancedMappingCrud.RichDomain.ServiceModel.Tests/AdvancedMappingCrud.RichDomain.ServiceModel.Tests.Domain/Entities/Product.cs
@@ -53,5 +53,19 @@ namespace AdvancedMappingCrud.RichDomain.ServiceModel.Tests.Domain.Entities
             // [IntentFully]
             throw new NotImplementedException("Replace with your implementation...");
         }
+
+        public async Task DoManualAsync(ICategoriesService catService, CancellationToken cancellationToken = default)
+        {
+            // [IntentFully]
+            // TODO: Implement DoManualAsync (Product) functionality
+            throw new NotImplementedException("Replace with your implementation...");
+        }
+
+        public void Another(ISecondService secondSevice)
+        {
+            // [IntentFully]
+            // TODO: Implement Another (Product) functionality
+            throw new NotImplementedException("Replace with your implementation...");
+        }
     }
 }