+++ Tests/AdvancedMappingCrud.RichDomain.Tests/AdvancedMappingCrud.RichDomain.Tests.Domain/Entities/Product.cs
@@ -57,5 +57,19 @@ namespace AdvancedMappingCrud.RichDomain.Tests.Domain.Entities
             var categories = categoryService.GetOrCreateCategoriesAsync(categoryNames).GetAwaiter().GetResult();
             Categories = categories.ToList();
         }
+
+        public async Task DoManualAsync(ICategoriesService catService, CancellationToken cancellationToken = default)
+        {
+            // [IntentFully]
+            // TODO: Implement DoManualAsync (Product) functionality
+            throw new NotImplementedException("Replace with your implementation...");
+        }
+
+        public void Another(ISecondService secondSevice)
+        {
+            // [IntentFully]
+            // TODO: Implement Another (Product) functionality
+            throw new NotImplementedException("Replace with your implementation...");
+        }
     }
 }