+++ Tests/AdvancedMappingCrud.RichDomain.Tests/AdvancedMappingCrud.RichDomain.Tests.Domain/Services/ISecondService.cs
@@ -1,0 +1,12 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.DomainServices.DomainServiceInterface", Version = "1.0")]
+
+namespace AdvancedMappingCrud.RichDomain.Tests.Domain.Services
+{
+    public interface ISecondService
+    {
+        void DoIt();
+    }
+}