+++ Tests/CleanArchitecture.Comprehensive/CleanArchitecture.Comprehensive.Api/Controllers/PaginationController.cs
@@ -5,6 +5,9 @@ using System.Threading.Tasks;
 using CleanArchitecture.Comprehensive.Application.Common.Pagination;
 using CleanArchitecture.Comprehensive.Application.Pagination;
 using CleanArchitecture.Comprehensive.Application.Pagination.GetLogEntries;
+using CleanArchitecture.Comprehensive.Application.Pagination.GetPeopleByFirstNamePaginated;
+using CleanArchitecture.Comprehensive.Application.Pagination.GetPeopleByNullableFirstNamePaginated;
+using CleanArchitecture.Comprehensive.Application.Pagination.GetPeoplePaginated;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
@@ -42,6 +45,62 @@ namespace CleanArchitecture.Comprehensive.Api.Controllers
             CancellationToken cancellationToken = default)
         {
             var result = await _mediator.Send(new GetLogEntriesQuery(pageNo: pageNo, pageSize: pageSize), cancellationToken);
+            return Ok(result);
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="200">Returns the specified PagedResult&lt;PersonEntryDto&gt;.</response>
+        /// <response code="400">One or more validation errors have occurred.</response>
+        [HttpGet("api/pagination/firstname/{firstName}")]
+        [ProducesResponseType(typeof(PagedResult<PersonEntryDto>), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<PagedResult<PersonEntryDto>>> GetPeopleByFirstNamePaginated(
+            [FromRoute] string firstName,
+            [FromQuery] int pageNo,
+            [FromQuery] int pageSize,
+            [FromQuery] string? orderBy,
+            CancellationToken cancellationToken = default)
+        {
+            var result = await _mediator.Send(new GetPeopleByFirstNamePaginatedQuery(firstName: firstName, pageNo: pageNo, pageSize: pageSize, orderBy: orderBy), cancellationToken);
+            return Ok(result);
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="200">Returns the specified PagedResult&lt;PersonEntryDto&gt;.</response>
+        /// <response code="400">One or more validation errors have occurred.</response>
+        [HttpGet("api/pagination/nullable-firstname/{firstName}")]
+        [ProducesResponseType(typeof(PagedResult<PersonEntryDto>), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<PagedResult<PersonEntryDto>>> GetPeopleByNullableFirstNamePaginated(
+            [FromRoute] string? firstName,
+            [FromQuery] int pageNo,
+            [FromQuery] int pageSize,
+            [FromQuery] string? orderBy,
+            CancellationToken cancellationToken = default)
+        {
+            var result = await _mediator.Send(new GetPeopleByNullableFirstNamePaginatedQuery(firstName: firstName, pageNo: pageNo, pageSize: pageSize, orderBy: orderBy), cancellationToken);
+            return Ok(result);
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="200">Returns the specified PagedResult&lt;PersonEntryDto&gt;.</response>
+        /// <response code="400">One or more validation errors have occurred.</response>
+        [HttpGet("api/pagination")]
+        [ProducesResponseType(typeof(PagedResult<PersonEntryDto>), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<PagedResult<PersonEntryDto>>> GetPeoplePaginated(
+            [FromQuery] int pageNo,
+            [FromQuery] int pageSize,
+            [FromQuery] string? orderBy,
+            CancellationToken cancellationToken = default)
+        {
+            var result = await _mediator.Send(new GetPeoplePaginatedQuery(pageNo: pageNo, pageSize: pageSize, orderBy: orderBy), cancellationToken);
             return Ok(result);
         }
     }